sstatesig: Allow the BB_HASHFILENAME to be passed in for comparision purposes
Currently the code can match too many things, for example, searching for
gcc-cross, we'd also get gcc-cross-initial.

If we have a bitbake cache available, we'd often have the BB_HASHFILENAME
data available from that cache. That can give us something to filter the
matches against.

The format of that field is metadata defined so it needs to be processed
in a metadata function (which find_siginfo is).

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>