meta:lib:oeqa:utils:qemurunner Add ability to launch qemu from python script
Add the ability to launch qemu using a <Machine>starter class which can
be included in oeqa/utils directory, either in the base meta layer or in
a bsp layer.
If no such class exists default back to using runqemu.
Pass the builds nativesysroot location so the <Machine>starter class
has access to the built qemu.
The <Machine>starter class will specify a ips and port numbers to use
for the ssh access. This allows the starter class the ability to redirect
ports if necessary.

Signed-off-by: Sipke Vriend <sipke.vriend@xilinx.com>